requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
python-docx
langchain
langchain-community
sentence-transformers
faiss-cpu